requirements.txt
absl-py==2.3.1
astunparse==1.6.3
async-timeout==5.0.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Cython==3.1.6
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.9.0
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==2.13.1
large-image==1.20.0
large-image-source-tiff==1.20.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
networkx==3.4.2
numpy==1.24.3
oauthlib==3.3.1
opencv-python-headless==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
palettable==3.3.3
pandas==2.3.3
pillow==12.0.0
protobuf==4.25.8
psutil==7.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pylibtiff==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==4.5.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tifftools==1.6.1
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
